// src/app/shared/app/models/Users/iusers.ts
export interface IUsers {
  sno?: number;
  clientId?: number;
  clientName?: string;
  employeeNo?: number;
  email?: string;
  password?: string;
  userType?: string;
  status?: string;
  title?: string;
  firstName?: string;
  surname?: string;
  position?: string;
  createdBy?: string;
  createdOn?: Date;
  lastUpdatedBy?: string;
  lastUpdatedOn?: Date;
}